+++ doc/每日进度文档.md
@@ -14,7 +14,7 @@
 
 ### 心得体会
 
-​	md我好像选了个最难的
+​	我好像选了个最难的
 
 
 
@@ -202,9 +202,7 @@
 
 ### 心得体会
 
-
-
-
+​	加油加油
 
 ## 4.27	周四	第十三天
