--- src/main/java/constants/RoleConstants.java
@@ -1,7 +1,0 @@
-package constants;
-
-public class RoleConstants {
-    public static final String ADMIN = "admin";
-    public static final String TRUE = "true";
-    public static final String FALSE = "false";
-}
