+++ src/main/resources/MyConnectionPool.properties
@@ -1,12 +1,10 @@
 
 
 # \u5E08\u5144\u672C\u5730URL\u4E0D\u662Flocalhost\uFF0C\u7AEF\u53E3\u4E0D\u662F3306\u7684\u8BDD\uFF0C\u4E5F\u8981\u6539\u6210\u81EA\u5DF1\u7684
-# test02\u662F\u6211\u672C\u5730\u7684\u6570\u636E\u5E93\u540D\uFF0C\u5E08\u5144\u8BB0\u5F97\u4FEE\u6539\u6210\u81EA\u5DF1\u5BFC\u5165sql\u811A\u672C\u7684\u6570\u636E\u5E93\u540D\uFF0C\u4E0D\u7136\u4F1A\u62A5\u9519
+# qgifinal\u662F\u6211\u672C\u5730\u7684\u6570\u636E\u5E93\u540D\uFF0C\u5E08\u5144\u8BB0\u5F97\u4FEE\u6539\u6210\u81EA\u5DF1\u5BFC\u5165sql\u811A\u672C\u7684\u6570\u636E\u5E93\u540D\uFF0C\u4E0D\u7136\u4F1A\u62A5\u9519
 # username\u548Cpassword\u4E5F\u5F97\u4FEE\u6539\u6210\u81EA\u5DF1\u7684\u6570\u636E\u5E93\u7528\u6237\u540D\u548C\u5BC6\u7801
 url=jdbc:mysql://localhost:3306/qgfinal?useServerPrepStmts=true
-#driverClassName=com.mysql.jdbc.Driver
 driverClassName=com.mysql.cj.jdbc.Driver
-
 username=root
 password=123456
 initialSize=10
