+++ src/main/webapp/css/login.css
@@ -1,0 +1,99 @@
+body {
+    /*background-color: #f2f2f2;*/
+    background: url("../resources/cyberpunk2077-3.jpg");
+    background-size: cover;
+    font-family: Arial, sans-serif;
+    color: #333;
+}
+
+.login-container {
+    width: 475px;
+    margin: 70px auto;
+    padding: 25px;
+    border-radius: 10px;
+
+    background-color: rgba(255, 255, 255, 0.7);
+    opacity: 0.87; /* 设置透明度 */
+    box-shadow: 0px 0px 5px #ccc;
+}
+
+h1 {
+    text-align: center;
+    font-size: 27px;
+    margin-bottom: 20px;
+}
+
+.form-group {
+    margin-bottom: 20px;
+}
+#rememberBtn{
+    display: inline-flex;
+    margin-left: 3px;
+}
+label {
+    display: block;
+    margin-bottom: 5px;
+    font-weight: bold;
+}
+#rememberBtnLabel{
+    display: inline-flex;
+    margin-bottom: 5px;
+    font-weight: bold;
+}
+input[type="text"],
+input[type="password"] {
+    display: inline;
+    width: 75%;
+    padding: 10px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+}
+
+button[type="submit"],
+button[type="button"]{
+    background-color: #4CAF50;
+    color: #fff;
+    border: none;
+    padding: 10px 20px;
+    margin-left: 5px;
+    margin-right: 5px;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+button[type="submit"],
+button[type="button"]:hover {
+    background-color: #3e8e41;
+}
+
+.error-message {
+    color: red;
+    /*text-align: center;*/
+    /*margin-top: 10px;*/
+    /*margin-top: 10px;*/
+    font-size: 13px;
+
+}
+#register-link {
+    display: inline-block;
+    margin-left: 8px;
+    /*margin-top: auto;*/
+    /*margin-bottom: auto;*/
+    /*font-size: 14px;*/
+    /*cursor: pointer;*/
+    color: #3d4bff;
+}
+#findPassword-link{
+    display: inline-block;
+    margin-left: 8px;
+    /*margin-top: auto;*/
+    /*margin-bottom: auto;*/
+    /*font-size: 14px;*/
+    /*cursor: pointer;*/
+    color: #426cff
+}
+#registerRoleClass{
+    display: flex;
+    justify-content: center;
+
+}
