+++ src/main/webapp/css/printerDetail.css
@@ -1,0 +1,79 @@
+/* 全局样式 */
+body {
+    font-family: Arial, sans-serif;
+    background-color: #f8f8f8;
+}
+
+/* 导航栏样式 */
+nav {
+    background-color: #333;
+    color: #fff;
+    height: 50px;
+    padding: 0 20px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+nav a {
+    color: #fff;
+    text-decoration: none;
+    font-size: 18px;
+    font-weight: bold;
+}
+
+/* 主要内容区样式 */
+main {
+    margin: 50px auto;
+    width: 80%;
+    max-width: 1200px;
+}
+
+h1 {
+    font-size: 36px;
+    text-align: center;
+    margin-bottom: 50px;
+}
+
+table {
+    width: 100%;
+    border-collapse: collapse;
+    margin-bottom: 30px;
+}
+
+/*table td {*/
+/*    padding: 10px 20px;*/
+/*    font-size: 18px;*/
+/*    border: 1px solid #ddd;*/
+/*}*/
+.staticTD{
+    padding: 12px 15px;
+    width: 32%;
+    text-align: center;
+    font-size: 17px;
+    border: 1px solid #ddd;
+}
+.dynamicTD{
+    padding: 15px 20px;
+    font-size: 17px;
+    border: 1px solid #ddd;
+}
+#title {
+    margin-top: 0;
+}
+.button {
+    display: inline-block;
+    padding: 8px 20px;
+    font-size: 16px;
+    font-weight: bold;
+    text-decoration: none;
+    color: #fff;
+    background-color: #333;
+    border-radius: 5px;
+    transition: all 0.3s ease;
+    margin-bottom: 15px;
+}
+
+.button:hover {
+    background-color: #555;
+}
