--- src/main/webapp/hall.css
@@ -1,96 +1,0 @@
-/* 全局样式 */
-body {
-    font-family: Arial, sans-serif;
-    background-color: #f8f8f8;
-}
-
-/* 导航栏样式 */
-nav {
-    background-color: #333;
-    color: #fff;
-    height: 50px;
-    padding: 0 20px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-nav a {
-    color: #fff;
-    text-decoration: none;
-    font-size: 18px;
-    font-weight: bold;
-}
-nav #userCenterBtn:hover {
-    text-decoration: underline;
-}
-nav #logoutBtn:hover {
-    text-decoration: underline;
-}
-/* 主要内容区样式 */
-main {
-    margin: 50px auto;
-    width: 80%;
-    max-width: 1200px;
-}
-
-h1 {
-    font-size: 36px;
-    text-align: center;
-    margin-bottom: 50px;
-}
-
-p {
-    font-size: 18px;
-    line-height: 1.5;
-    text-align: justify;
-    margin-bottom: 20px;
-}
-
-.printer-link {
-    display: block;
-    padding: 8px;
-    text-align: center;
-    margin: 10px 0;
-    font-size: 18px;
-    text-decoration: none;
-    color: #c612ff;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    transition: all 0.3s ease;
-}
-
-.printer-link:hover {
-    background-color: #f8f8f8;
-    border-color: #333;
-    color: #333;
-    text-decoration: none;
-}
-.button {
-    display: inline-block;
-    padding: 8px 20px;
-    font-size: 16px;
-    font-weight: bold;
-    text-decoration: none;
-    color: #fff;
-    background-color: #333;
-    border-radius: 5px;
-    transition: all 0.3s ease;
-}
-
-.button:hover {
-    background-color: #555;
-}
-
-/* 页脚样式 */
-footer {
-    position: absolute;
-    text-align: center;
-    width: 100%;
-    margin-top: 50px;
-    bottom: 0;
-    padding: 20px;
-    background-color: #333;
-    color: #fff;
-}
-
