+++ src/main/webapp/hall.html
@@ -10,10 +10,10 @@
 <head>
   <meta charset="UTF-8">
   <title>登录后页面</title>
-  <link rel="stylesheet" href="hall.css">
+  <link rel="stylesheet" href="css/hall.css">
   <script src="https://lf3-cdn-tos.bytecdntp.com/cdn/expire-1-M/jquery/3.6.0/jquery.min.js"></script>
 
-  <script src="hall.js"></script>
+  <script src="webjars/hall.js"></script>
 </head>
 <body>
 <!-- 导航栏 -->
