+++ src/main/webapp/login.html
@@ -3,9 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <title>Login</title>
-    <link rel="stylesheet" href="login.css">
+    <link rel="stylesheet" href="css/login.css">
     <script src="https://lf3-cdn-tos.bytecdntp.com/cdn/expire-1-M/jquery/3.6.0/jquery.min.js"></script>
-    <script src="login.js"></script>
+    <script src="webjars/login.js"></script>
 </head>
 <body>
 <div class="login-container">
@@ -24,7 +24,6 @@
             <label for="password">密码:</label>
             <input type="password" id="password" name="password" placeholder="请输入密码" required>
             <span class="register">
-            <a href="login.js" id="findPassword-link">忘记密码?</a>
             </span>
             <p>
                 <label for="rememberBtn" id="rememberBtnLabel">记住我:</label>
