+++ src/main/webapp/printerDetail.html
@@ -3,9 +3,9 @@
 <head>
   <meta charset="UTF-8">
   <title>Printer Details</title>
-  <link rel="stylesheet" href="printerDetail.css">
+  <link rel="stylesheet" href="css/printerDetail.css">
   <script src="https://lf3-cdn-tos.bytecdntp.com/cdn/expire-1-M/jquery/3.6.0/jquery.min.js"></script>
-  <script src="printerDetail.js"></script>
+  <script src="webjars/printerDetail.js"></script>
 </head>
 <body>
 <!-- 导航栏 -->
@@ -17,7 +17,7 @@
 <!-- 主要内容区 -->
 <main>
   <h1 id="title">打印机详情信息</h1>
-  <button class="button" id="startBtn">启动后台监听线程</button>
+  <button class="button" id="startBtn" >启动后台监听线程</button>
   <table>
     <tr>
       <td class="staticTD">打印机名：</td>
