+++ src/main/webapp/register.html
@@ -3,10 +3,10 @@
 <head>
     <meta charset="UTF-8">
     <title>register</title>
-    <link rel="stylesheet" href="login.css">
+    <link rel="stylesheet" href="css/login.css">
     <script src="https://lf3-cdn-tos.bytecdntp.com/cdn/expire-1-M/jquery/3.6.0/jquery.min.js"
             type="application/javascript"></script>
-    <script src="register.js" type="application/javascript"></script>
+    <script src="webjars/register.js" type="application/javascript"></script>
 </head>
 <body>
 <div class="login-container">
