--- src/main/webapp/test.html
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Printer Details</title>
-    <link rel="stylesheet" href="test.css">
-    <script src="test.js">
-    <script src="https://lf3-cdn-tos.bytecdntp.com/cdn/expire-1-M/jquery/3.6.0/jquery.min.js"></script>
-</head>
-<body>
-<!-- 导航栏 -->
-<nav>
-    <a>打印机管理系统</a>
-    <a href="#" id="backBtn">返回</a>
-</nav>
-
-<!-- 主要内容区 -->
-<main>
-    <h1 id="title">打印机详情信息</h1>
-    <button class="button" id="startBtn">启动后台监听线程</button>
-    <table>
-        <tr>
-            <td>打印机状态：</td>
-            <td id="printerStatus"></td>
-        </tr>
-        <tr>
-            <td>当前状态描述：</td>
-            <td id="printerReason"></td>
-        </tr>
-        <tr>
-            <td>任务进度：</td>
-            <td id="taskProgress"></td>
-        </tr>
-        <tr>
-            <td>当前温度：</td>
-            <td id="printerTemperature"></td>
-        </tr>
-        <tr>
-            <td>今日累积开机时间：</td>
-            <td id="uptimeToday"></td>
-        </tr>
-        <tr>
-            <td>今日累积打印时间：</td>
-            <td id="printTimeToday"></td>
-        </tr>
-        <tr>
-            <td>今日累积闲置时间：</td>
-            <td id="idleTimeToday"></td>
-        </tr>
-        <tr>
-            <td>今日累积异常时间：</td>
-            <td id="errorTimeToday"></td>
-        </tr>
-    </table>
-
-    <button id="refreshBtn">刷新</button>
-</main>
-
-<!-- 页脚 -->
-<footer>
-    <div id="currentTime"></div>
-</footer>
-
-<script src="test.js"></script>
-</body>
-</html>
