+++ src/main/webapp/userCenter.html
@@ -3,8 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <script src="https://lf3-cdn-tos.bytecdntp.com/cdn/expire-1-M/jquery/3.6.0/jquery.min.js"></script>
-    <link rel="stylesheet" href="userCenter.css">
-    <script src="userCenter.js"></script>
+    <link rel="stylesheet" href="css/userCenter.css">
+    <script src="webjars/userCenter.js"></script>
     <title>userCenter</title>
 </head>
 <body>
